+++ noodles-csi/src/reader.rs
@@ -142,7 +142,9 @@ where
     let mut metadata = None;
 
     for _ in 0..n_bin {
-        let id = reader.read_u32::<LittleEndian>()?;
+        let id = reader.read_u32::<LittleEndian>().and_then(|n| {
+            usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        })?;
 
         let loffset = reader
             .read_u64::<LittleEndian>()